refactor component imports in favour of the new phoenix component strucutre
